+++ js/game.js
@@ -1,0 +1,1 @@
+// сама игра
