+++ js/setup.js
@@ -1,0 +1,2 @@
+//настройка игры
+
